Change gid from int64_t to size_t

+++ shared/include/shared/input.h
@@ -1,10 +1,12 @@
 #ifndef INPUT_H
 #define INPUT_H
 
+#include <stdint.h>
+
 struct input
 {
-    int dx;
-    int dy;
+    int8_t dx;
+    int8_t dy;
 };
 
 #endif
